Just noticed a potential dupe glitch.
Whenever a player enters a new dimension, we save their inventory from the previous dimension. The potential dupe is that if a player enters a new dimension, we save their inventory, then they die and drop their items in that dimension. When they return to that dimension to retrieve their items, we would load their inventory again as it was on the first trip, thus duping the items they dropped.